Issue #248 RectMapCollider, player sometimes stuck
See https://github.com/los-cocos/cocos/issues/248
The problem is that the collision resolution is applied in any
order. Now we only move the rect by the minimum distance. If we
need to move on both axis, we leave this cell for later checking,
resolving first all the other collisions.