Fix TmxObjectMapCollider, #252
Objects to be considered in a TmxObjectLayer do not need a tile.
In order to have a similar collide_map method with RectMapCollider
the change has been factored out in a new method
get_cells_in_region. It takes the map and the rect to return the
set of objects or rects to consider in the collide_map method.